refactor(api): extract no-new-keyword logic from constructor to an extension and apply at entry point
