+++ src/extensions.js
@@ -1,0 +1,1 @@
+export * from './extensions/withNoNewKeyword'
